Move various processing to the parse loop
When parsing the config file we can resolve a number of references, like
the component variables to make the code simpler
