+++ internal/config/parse.go
@@ -4,14 +4,11 @@ import (
 	"context"
 	"errors"
 	"fmt"
-	"os"
 	"path/filepath"
-	"regexp"
 	"strconv"
 	"strings"
 
 	"github.com/creasty/defaults"
-	"github.com/elliotchance/pie/v2"
 	"gopkg.in/yaml.v3"
 
 	"github.com/labd/mach-composer/internal/plugins"
@@ -48,7 +45,7 @@ func Load(ctx context.Context, filename string, varFilename string) (*MachConfig
 		return nil, fmt.Errorf("failed to load config %s due to errors", filename)
 	}
 
-	cfg, err := parseConfig(ctx, body, vars, filename)
+	cfg, err := ParseConfig(ctx, body, vars, filename)
 	if err != nil {
 		return nil, err
 	}
@@ -58,10 +55,6 @@ func Load(ctx context.Context, filename string, varFilename string) (*MachConfig
 	}
 
 	cfg.Filename = filepath.Base(filename)
-	if err := ProcessConfig(cfg); err != nil {
-		return nil, err
-	}
-
 	return cfg, nil
 }
 
@@ -92,7 +85,7 @@ func getSchemaVersion(data []byte) (int, error) {
 
 // parseConfig is responsible for parsing a mach composer yaml config file and
 // creating the resulting MachConfig struct.
-func parseConfig(ctx context.Context, data []byte, vars *variables.Variables, filename string) (*MachConfig, error) {
+func ParseConfig(ctx context.Context, data []byte, vars *variables.Variables, filename string) (*MachConfig, error) {
 	// Decode the yaml in an intermediate config file
 	intermediate := &_RawMachConfig{}
 	err := yaml.Unmarshal(data, intermediate)
@@ -128,7 +121,7 @@ func parseConfig(ctx context.Context, data []byte, vars *variables.Variables, fi
 	}
 
 	if vars.Encrypted {
-		err := addFileToConfig(cfg, intermediate.MachComposer.VariablesFile)
+		err := cfg.addFileToConfig(intermediate.MachComposer.VariablesFile)
 		if err != nil {
 			return nil, err
 		}
@@ -138,12 +131,12 @@ func parseConfig(ctx context.Context, data []byte, vars *variables.Variables, fi
 		return nil, fmt.Errorf("failed to parse global node: %w", err)
 	}
 
-	if err := parseSitesNode(cfg, &intermediate.Sites); err != nil {
-		return nil, fmt.Errorf("failed to parse sites node: %w", err)
-	}
-
 	if err := parseComponentsNode(cfg, &intermediate.Components, filename); err != nil {
 		return nil, fmt.Errorf("failed to parse components node: %w", err)
+	}
+
+	if err := parseSitesNode(cfg, &intermediate.Sites); err != nil {
+		return nil, fmt.Errorf("failed to parse sites node: %w", err)
 	}
 
 	return cfg, nil
@@ -156,18 +149,12 @@ func parseGlobalNode(cfg *MachConfig, globalNode *yaml.Node) error {
 
 	knownKeys := []string{"cloud", "terraform_config", "environment"}
 	nodes := mapYamlNodes(globalNode.Content)
-	for key, node := range nodes {
-		if pie.Contains(knownKeys, key) {
-			continue
-		}
 
-		data, err := nodeAsMap(node)
-		if err != nil {
-			return err
-		}
-		if err := cfg.Plugins.SetGlobalConfig(key, data); err != nil {
-			return err
-		}
+	err := iterateYamlNodes(nodes, knownKeys, func(key string, data map[string]any) error {
+		return cfg.Plugins.SetGlobalConfig(key, data)
+	})
+	if err != nil {
+		return err
 	}
 
 	if node, ok := nodes["terraform_config"]; ok {
@@ -199,222 +186,12 @@ func parseGlobalNode(cfg *MachConfig, globalNode *yaml.Node) error {
 	return nil
 }
 
-func parseSitesNode(cfg *MachConfig, sitesNode *yaml.Node) error {
-	if err := sitesNode.Decode(&cfg.Sites); err != nil {
-		return fmt.Errorf("decoding error: %w", err)
-	}
-
-	cloudPlugin, err := cfg.Plugins.Get(cfg.Global.Cloud)
-	if err != nil {
-		return err
-	}
-
-	knownKeys := []string{
-		"name", "identifier", "endpoints", "components",
-	}
-	for _, site := range sitesNode.Content {
-		nodes := mapYamlNodes(site.Content)
-		siteId := nodes["identifier"].Value
-
-		for key, node := range nodes {
-			if pie.Contains(knownKeys, key) {
-				continue
-			}
-			data, err := nodeAsMap(node)
-			if err != nil {
-				return err
-			}
-
-			if err := cfg.Plugins.SetSiteConfig(key, siteId, data); err != nil {
-				return err
-			}
-		}
-
-		if node, ok := nodes["endpoints"]; ok {
-			data, err := nodeAsMap(node)
-			if err != nil {
-				return err
-			}
-			if err := cloudPlugin.SetSiteEndpointsConfig(siteId, data); err != nil {
-				return err
-			}
-		}
-
-		if err := parseSiteComponentsNode(cfg, siteId, nodes["components"]); err != nil {
-			return err
-		}
-	}
-
-	return nil
-}
-
-func parseSiteComponentsNode(cfg *MachConfig, site string, node *yaml.Node) error {
-	knownKeys := []string{
-		"name", "variables", "secrets",
-		"store_variables", "store_secrets",
-	}
-	for _, component := range node.Content {
-		nodes := mapYamlNodes(component.Content)
-		identifier := nodes["name"].Value
-		migrateCommercetools(identifier, nodes)
-
-		for key, node := range nodes {
-			if pie.Contains(knownKeys, key) {
-				continue
-			}
-			data, err := nodeAsMap(node)
-			if err != nil {
-				return err
-			}
-			if err := cfg.Plugins.SetSiteComponentConfig(site, identifier, key, data); err != nil {
-				return err
-			}
-		}
-	}
-	return nil
-}
-
-func addFileToConfig(cfg *MachConfig, filename string) error {
-	b, err := os.ReadFile(filename)
-	if err != nil {
-		return fmt.Errorf("error reading variables file: %w", err)
-	}
-	filename = filepath.Base(filename)
-	cfg.ExtraFiles[filename] = b
-	return nil
-}
-
-func parseComponentsNode(cfg *MachConfig, node *yaml.Node, source string) error {
-	if node.Tag == "!!str" {
-		path := filepath.Dir(source)
-		var err error
-		node, err = loadComponentsNode(node, path)
-		if err != nil {
-			return err
-		}
-	}
-
-	if err := node.Decode(&cfg.Components); err != nil {
-		return fmt.Errorf("decoding error: %w", err)
-	}
-
-	cloudPlugin, err := cfg.Plugins.Get(cfg.Global.Cloud)
-	if err != nil {
-		return err
-	}
-	for i := range cfg.Components {
-		c := &cfg.Components[i]
-		err := cloudPlugin.SetComponentEndpointsConfig(c.Name, c.Endpoints)
-		if err != nil {
-			return err
-		}
-	}
-
-	knownKeys := []string{
-		"name", "source", "version", "branch",
-		"integrations", "endpoints",
-	}
-	for _, component := range node.Content {
-		nodes := mapYamlNodes(component.Content)
-		identifier := nodes["name"].Value
-
-		for key, node := range nodes {
-			if pie.Contains(knownKeys, key) {
-				continue
-			}
-
-			data, err := nodeAsMap(node)
-			if err != nil {
-				return err
-			}
-
-			if err := cfg.Plugins.SetComponentConfig(key, identifier, data); err != nil {
-				return err
-			}
-		}
-	}
-
-	return nil
-}
-
-func loadComponentsNode(node *yaml.Node, path string) (*yaml.Node, error) {
-	re := regexp.MustCompile(`\$\{include\(([^)]+)\)\}`)
-	data := re.FindStringSubmatch(node.Value)
-	if len(data) != 2 {
-		return nil, fmt.Errorf("failed to parse ${include()} tag")
-	}
-	filename := filepath.Join(path, data[1])
-	body, err := utils.AFS.ReadFile(filename)
-	if err != nil {
-		return nil, err
-	}
-
-	result := yaml.Node{}
-	if err = yaml.Unmarshal(body, &result); err != nil {
-		return nil, err
-	}
-	if len(result.Content) != 1 {
-		return nil, fmt.Errorf("Invalid yaml file")
-	}
-	return result.Content[0], nil
-}
-
 func nodeAsMap(n *yaml.Node) (map[string]any, error) {
 	target := map[string]any{}
 	if err := n.Decode(&target); err != nil {
 		return nil, err
 	}
 	return target, nil
-}
-
-// migrateCommercetools moves the store_variables and store_secrets under the
-// commercetools node. Needed to say backwards compatible
-func migrateCommercetools(name string, nodes map[string]*yaml.Node) {
-	needsMigrate := false
-	if _, ok := nodes["store_variables"]; ok {
-		needsMigrate = true
-	}
-	if _, ok := nodes["store_secrets"]; ok {
-		needsMigrate = true
-	}
-	if !needsMigrate {
-		return
-	}
-
-	fmt.Printf("Warning: %s move store_variables and store_secrets to commercetools node\n", name)
-
-	if _, ok := nodes["commercetools"]; !ok {
-		nodes["commercetools"] = &yaml.Node{
-			Kind:    yaml.MappingNode,
-			Tag:     "!!map",
-			Content: []*yaml.Node{},
-		}
-
-		if val, ok := nodes["store_variables"]; ok {
-			keyNode := &yaml.Node{
-				Kind:  yaml.ScalarNode,
-				Tag:   "!!str",
-				Value: "store_variables",
-			}
-			nodes["commercetools"].Content = append(
-				nodes["commercetools"].Content,
-				keyNode,
-				val,
-			)
-		}
-		if val, ok := nodes["store_secrets"]; ok {
-			keyNode := &yaml.Node{
-				Kind:  yaml.ScalarNode,
-				Tag:   "!!str",
-				Value: "store_secrets",
-			}
-			nodes["commercetools"].Content = append(
-				nodes["commercetools"].Content,
-				keyNode,
-				val,
-			)
-		}
-	}
 }
 
 func loadDefaultPlugins(r *plugins.PluginRepository) {
