+++ internal/generator/generate_test.go
@@ -1,67 +1,55 @@
 package generator
 
 import (
+	"context"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/stretchr/testify/require"
 
 	"github.com/labd/mach-composer/internal/config"
-	"github.com/labd/mach-composer/internal/plugins"
+	"github.com/labd/mach-composer/internal/utils"
 )
 
 func TestRenderSite(t *testing.T) {
-	cfg := config.MachConfig{
-		MachComposer: config.MachComposer{
-			Version: "1.0.0",
-		},
-		Global: config.Global{
-			Environment:            "test",
-			Cloud:                  "aws",
-			TerraformStateProvider: "aws",
-		},
-		Sites: []config.Site{
-			{
-				Name:       "",
-				Identifier: "my-site",
-				RawEndpoints: map[string]any{
-					"main": "api.my-site.nl",
-					"internal": map[string]any{
-						"throttling_burst_limit": 5000,
-						"throttling_rate_limit":  10000,
-						"url":                    "internal-api.my-site.nl",
-					},
-				},
-				Components: []config.SiteComponent{
-					{
-						Name: "your-component",
-						Variables: map[string]any{
-							"FOO_VAR": "my-value",
-						},
-						Secrets: map[string]any{
-							"MY_SECRET": "secretvalue",
-						},
-					},
-				},
-			},
-		},
-		Components: []config.Component{
-			{
-				Name:         "your-component",
-				Source:       "git::https://github.com/<username>/<your-component>.git//terraform",
-				Version:      "0.1.0",
-				Integrations: []string{"aws", "commercetools"},
-			},
-		},
-		Plugins: plugins.NewPluginRepository(),
+	data := []byte(utils.TrimIndent(`
+	---
+	mach_composer:
+	  version: 1.0.0
+	global:
+	  environment: test
+	  cloud: aws
+	  terraform_config:
+		aws_remote_state:
+	sites:
+	- identifier: my-site
+	  endpoints:
+		main: api.my-site.nl
+		internal:
+		  url: internal-api.my-site.nl
+		  throttling_burst_limit: 5000
+		  throttling_rate_limit: 10000
+	  components:
+	  - name: your-component
+		variables:
+		  FOO_VAR: my-value
+		secrets:
+		  MY_SECRET: secretvalue
+	components:
+	- name: your-component
+	  source: "git::https://github.com/<username>/<your-component>.git//terraform"
+	  version: 0.1.0
+	  endpoints:
+		internal: internal
+	  integrations:
+		- aws
+		- commercetools
+`))
+	cfg, err := config.ParseConfig(context.Background(), data, nil, "main.yml")
+	if err != nil {
+		t.Error(err)
 	}
-	err := cfg.Plugins.Load("aws", "internal")
-	require.NoError(t, err)
 
-	err = config.ProcessConfig(&cfg)
-	require.NoError(t, err)
-
-	body, err := renderSite(&cfg, &cfg.Sites[0])
+	body, err := renderSite(cfg, &cfg.Sites[0])
 	assert.NoError(t, err)
 	assert.NotEmpty(t, body)
 }
