+++ internal/updater/writer_config.go
@@ -5,9 +5,10 @@ import (
 	"log"
 	"strings"
 
-	"github.com/labd/mach-composer/internal/utils"
 	"github.com/sirupsen/logrus"
 	"gopkg.in/yaml.v3"
+
+	"github.com/labd/mach-composer/internal/utils"
 )
 
 type MachConfig struct {
